+++ src/main/java/com/creditmodule/ing/data/CreateOrderRequest.java
@@ -3,11 +3,13 @@ package com.creditmodule.ing.data;
 import com.creditmodule.ing.enums.Side;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateOrderRequest {
     private Long customerId;
     private String assetName;
     private Side side;
-    private double size;
+    private BigDecimal size;
 }
